Improved README.md
While "Floyd-Steinberg" is indeed of an algorithm, which can be considered a "thing", we must acknowledge that the name is a conjunction of proper nouns, specifically surnames. Based on this, I propose that we cease treating the phrase `Steinberg` as an unknown object currently designated by a "What?", but as person unfamiliar to the reader. Hence, within this commit, I am proposing the changing of "What?" to "Who?".